+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 * [HexaPDF::Revision#each] to allow iteration over loaded objects only
 * [HexaPDF::Document#each] method argument from `current` to `only_current`
+* [HexaPDF::Object#==] and [HexaPDF::Reference#==] so that Object and Reference
+  objects can be compared
 
 ### Fixed